% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS.R
\name{check_phenotype}
\alias{check_phenotype}
\title{Check a specified phenotype}
\usage{
check_phenotype(phenotype)
}
\arguments{
\item{phenotype}{Either a character vector of length one (i.e. a name) or
an object inheriting from class \code{FGRS_data}}
}
\value{
An object of class \code{FGRS_data}
}
\description{
Given either a name or an object, this function checks whether a built-in
data object for the name exists, or whether the object inherits from a
valid phenotype structure. If yes, the correct phenotype is returned,
otherwise an error is thrown.
}
\seealso{
\code{\link[FGRS]{FGRS}} \code{\link[FGRS]{check_probands}} \code{\link[FGRS]{FGRS_data}}
}
