% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS.R
\name{check_probands}
\alias{check_probands}
\alias{check_relatives}
\title{Check a data frame of probands or relatives for correctness}
\usage{
check_probands(x, id = "ProbandID", byear = "BirthYear")

check_relatives(
  x,
  id = "ProbandID",
  id2 = "RelativeID",
  sex = "Sex",
  byear = "BirthYear",
  dyear = "DiagYear",
  age = "AgeEOF",
  reltyp = "RelType",
  gensh = "SharedGenetics"
)
}
\description{
Given data frames that contain information on either probands or relatives
as required by function \code{FGRS}, this function runs some minimal checks on
correctness and returns the subset of coumns required to calculate the FGRS.
}
\seealso{
\code{\link[FGRS]{FGRS}}
}
