% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS.R
\name{cumulativeFGRS}
\alias{cumulativeFGRS}
\title{Calculate partial family genetic risk scores}
\usage{
cumulativeFGRS(probands, relatives, phenotype)
}
\arguments{
\item{probands}{Proband information, same as for \code{FGRS}}

\item{relatives}{Information on relatives; same as for \code{FGRS}}

\item{phenotype}{An object of class \code{FGRS_data} containing the marginal
population data required to calculate the FGRS, same as for
\code{FGRS}.}
}
\value{
A data frame with either eight or seven columns, containing proband
IDs, weighted nubmer of relatives and six variant FGRS scores

\if{html}{\out{<div class="sourceCode">}}\preformatted{    * `ProbandID`: unique identifier for probands, as specified in argument
                  `probands`
    * `nRelatives`: number of relatives weighted by shared genetics
    * Either six or five variant FGRS scores, called either `cumFGRS0` to
      `cumFGRS5` or `looFGRS1` to `looFGRS5`, depending on function;
       see Details.

     Probands without relatives will lead to a warning and a missing value
     for the corresponding FGRS. Relatives unrelated to any specified
     proband will lead to a warning and will be dropped.

    This data frame does _not_ carry information on intermediate results
    on the relative level, as in `FGRS`
}\if{html}{\out{</div>}}
}
\description{
Calculate variant FGRS by skipping one or several steps of the full FGRS
algorithm: \code{cumulativeGFRS} calculates partial FGRS by performing steps in
order, but terminating before the end; \code{leaveoneoutFGRS} calculates incomplete
FGRS by skipping each step of the algorithm in turn. Both functions return
a series of FGRS variants that can be used to study the properties of the
full FGRS, e.g. as a sensitivity analysis.
}
\details{
For the purpose of these functions, the FGRS is calculated in six steps,
numbered 0-5:

\if{html}{\out{<div class="sourceCode">}}\preformatted{     0. Calculation of raw liabilities based on sex and birth decade
        of relative
     1. Age correction based on sex, age and disease status at end of
        follow-up (in relatives)
     2. Correction for genetic relatedness to proband (in relatives)
     3. Correction for cohabitation (in relatives)
     4. Shrinkage based on family size (in probands)
     5. Standardization based on birth year (in probands)

     Note that this is not the same order of steps as in the reference
     below, but makes more sense like this here.

     Note that the algorithm includes an averaging step between Steps 3
     and 4 above when going from relatives to probands. This is necessary
     and will always be run.

     `cumFGRS0` represents the raw liability for disease among relatives;
     `cumFGRS1` to `cumFGRS5` represent the output from the algorithm after
     running the 1st to 5th step of the algorithm - consequently,
     `cumFGRS5` is just the full FGRS again.

     `looFGRS1` to `looFGRS5` represent the output from the FGRS algorithm
     when skipping the the 1st to 5th step.
}\if{html}{\out{</div>}}
}
\examples{
sz_cum <- cumulativeFGRS(ex1$probands, ex1$relatives, phenotype = "SZ")
sz_cum

sz_loo <- leaveoneoutFGRS(ex1$probands, ex1$relatives, phenotype = "SZ")
sz_loo
}
\references{
Kendler, K., Ohlsson, H., Sundquist, J., & Sundquist, K. (2021).
Impact of comorbidity on family genetic risk profiles for psychiatric and
substance use disorders: A descriptive analysis. Psychological Medicine, 1-10.
doi:10.1017/S0033291721004268
}
\seealso{
\code{\link[FGRS]{FGRS}}
}
