% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{Example data for FGRS calculation}
\format{
A list with two data frames, \code{probands} and \code{relatives}
}
\usage{
ex1
}
\description{
A small synthetic data set that tracks the same diagnosis in both
probands and relatives
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{FIXME
}\if{html}{\out{</div>}}
}
\keyword{datasets}
