% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS_classes.R
\name{FGRS_data}
\alias{FGRS_data}
\title{Population data required for the Kendler-Ohlsson family genetic risk score}
\description{
A class that captures all population data required to calculate
the Kendler-Ohlsson FGRS for members of the underlying population
}
\references{
Kendler, K. S., Ohlsson, H., Sundquist, J., & Sundquist, K. (2021a).
Family genetic risk scores and the genetic architecture of major affective and
psychotic disorders in a Swedish national sample. JAMA Psychiatry, 78, 735–743.

Kendler, K., Ohlsson, H., Sundquist, J., & Sundquist, K. (2021b).
Impact of comorbidity on family genetic risk profiles for psychiatric and
substance use disorders: A descriptive analysis. Psychological Medicine, 1-10.
doi:10.1017/S0033291721004268
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{disease}}{A character vector of length one, containing the name of the
disease for which the FGRS data has been collected}

\item{\code{birth_decades}}{A data frame with three numerical columns, mapping the
codes used for birth decades to actual ranges of birth years:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `Start`: starting year of birth decade (excluded from decade)
   * `End`: end year of birth decade (included in decade)
   * `BirthDecade`: numerical code for birth decade
}\if{html}{\out{</div>}}}

\item{\code{age_case_cuminc}}{Data frame with three numerical columns, containing
the cumulative incidence by age at first diagnosis for all diagnosed
cases of the disease in the underlying population, separately for
males and females; see Kendler et al. 2021(b), Supplemental Table 2,
Step 1 (note: cumulative incidences stored here are \emph{not} adjusted
for registration effects).

\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `Age`: age at diagnosis
   * `PropDiag`: proportion of cases diagnosed before or at the
                 corresponding age
   * `Sex`: indicator for sex (1 = male, 2 = female)
}\if{html}{\out{</div>}}}

\item{\code{bdecade_liability}}{Data frame with six numerical columns, listing
mean libability for the disease of interest by birth decade, sex and
disease status at end of follow-up; ; see Kendler et al. 2021(b),
Supplemental Table 2, Step 2.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `BirthDecade`: numerical code for birth decade, as defined by
                    field `BirthDecade`
   * `Sex`: indicator for sex (1 = male, 2 = female)
   * `HasDiag`: indicator for disease status at end of follow-up (0 =
                no diagnosis, 1 = diagnosed case)
   * `MeanLiab`: the estimated mean liability for disease for the
                 recorded combination of borth decade, sex and disease
                 status (via a truncated normal distribution)
   * `LiabThresh`: the liability threshold between undiagnosed subjects
                   and diagnosed subjects, based on the observed
                   prevalence; uniquely defines mean liability for cases
                   and controls, but only included for description.
   * `Prev`: the observed prevalence of the disease for subjects with
             the recorded combination of birth decade and sex; uniquely
             defines mean liability for cases and controls, but only
             included for description.
}\if{html}{\out{</div>}}}

\item{\code{cohab_parent_child, cohab_sibling}}{Numerical vectors of lenght one,
storing the correction factors for cohabitation effects (shared
environment) for parent-child and sibling-sibling relationships;
see Kendler et al. 2021(b), Supplemental Table 2, Step 3.}

\item{\code{shrinkage_A, shrinkage_B}}{Numerical vectors of length one,
holding the correction factors for the number of relatives (family
size) recorded for a proband; see Kendler et al. 2021(b),
Supplemental Table 2, Step 6.}

\item{\code{byear_standard}}{Data frame with three numerical columns, for
standardizing the calculated FGRS values by birth year of the
proband; see Kendler et al. 2021(b), Supplemental Table 2, Step 7.
Note: this implementation does not include standardisation by county
of longest residence.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `Byear`: birth year for the borth cohort
   * `Mean`: mean un-standardized FGRS across the birth cohort
   * `STD`: standard deviation of un-standardized FGRS across the birth cohort
}\if{html}{\out{</div>}}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohab_parent_child, cohab_sibling}}{Numerical vectors of lenght one,
storing the correction factors for cohabitation effects (shared
environment) for parent-child and sibling-sibling relationships;
see Kendler et al. 2021(b), Supplemental Table 2, Step 3.}

\item{\code{shrinkage_A, shrinkage_B}}{Numerical vectors of length one,
holding the correction factors for the number of relatives (family
size) recorded for a proband; see Kendler et al. 2021(b),
Supplemental Table 2, Step 6.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FGRS_data-new}{\code{FGRS_data$new()}}
\item \href{#method-FGRS_data-get_age_weights}{\code{FGRS_data$get_age_weights()}}
\item \href{#method-FGRS_data-get_disease_liability}{\code{FGRS_data$get_disease_liability()}}
\item \href{#method-FGRS_data-get_cohab_corrfac}{\code{FGRS_data$get_cohab_corrfac()}}
\item \href{#method-FGRS_data-get_famsize_corrfac}{\code{FGRS_data$get_famsize_corrfac()}}
\item \href{#method-FGRS_data-standardize_fgrs}{\code{FGRS_data$standardize_fgrs()}}
\item \href{#method-FGRS_data-clone}{\code{FGRS_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-new}{}}}
\subsection{Method \code{new()}}{
Initialize an empty \code{FGRS_data} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$new(
  name,
  birth_decades,
  age_case_cuminc,
  bdecade_liability,
  cohab,
  shrinkage,
  byear_standard
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character vector of length one, the name of the phenotype for
which the FGRS data was collected}

\item{\code{birth_decades}}{Data frame with three numerical columns, coding for the
birth decades used in the data: \code{Start}, \code{End},
\code{BirthDecade}. Intervals are open on the left and
closed on the right-}

\item{\code{age_case_cuminc}}{Data frame with three numerical columns, listing
cumulative incidence of the diagnosis of interest
among cases, by age and sex: \code{Age}, \code{PropDiag} and
\code{Sex} (where 1 = male and 2 = female)}

\item{\code{bdecade_liability}}{Data frame with three numerical columns, listing the liability
threshold separating subjects with and without
diagnosis, by birth decade and sex: \code{BirthDecade}
(numerical code as for argument \code{birth_decades}),
\code{Sex}(where 1 = male and 2 = female) and \code{LiabThresh}
(the actual liability threshold)}

\item{\code{cohab}}{A named numerical vector or length two, listing the correction
factors for cohabitation between parents and children
(\code{parent_child}) and between siblings (\code{siblings})}

\item{\code{shrinkage}}{A named numerical vector or length two, listing the variance
components that define shrinkage factor accounting for the
number of relatives recorded for a proband (names:\code{A} and
\code{B}, as in Kendler et al. 2021b, Supp.Table 2, Step 6)}

\item{\code{byear_standard}}{A data frame with three columns, listing mean and
standard deviation of the unstandardized FGRS by
birth year of the probands: \code{Byear}, \code{Mean}, \code{STD}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function does some hard checks for consistency which will lead
to errors if violated
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-get_age_weights"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-get_age_weights}{}}}
\subsection{Method \code{get_age_weights()}}{
Age weights for relatives without diagnosis

Relatives that have not received a diagnosis by end of follow-up will be
weighted based on the cumulative incidence of diagnosis among cases at their
age at end of follow-up: in other words, a relative still un-diagnosed at
older age will get more weight than one at younger age. Relatives with
a diagnosis will always get full weight.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$get_age_weights(byear, sex, age_eof, has_diag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{byear}}{Numeric vector, year of birth of relative}

\item{\code{sex}}{Integer vector, indicating the sex of the relative (1 = male,
2 = female)}

\item{\code{age_eof}}{Numeric vector, age of relative at end of follow-up}

\item{\code{has_diag}}{Logical vector, indicating whether the relative has a
diagnosis at end of follow-up}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fractional ages are currently linearly interpolated, but this may
change
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-get_disease_liability"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-get_disease_liability}{}}}
\subsection{Method \code{get_disease_liability()}}{
Mean liabilities of relatives with and without diagnosis

Relatives are assigned an underlying libabilitiy for the phenotype of
interest, based on its prevalence by birth decade and sex
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$get_disease_liability(byear, sex, has_diag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{byear}}{Numeric vector, year of birth of relative}

\item{\code{sex}}{Integer vector, indicating the sex of the relative (1 = male,
2 = female)}

\item{\code{has_diag}}{Logical vector, indicating whether the relative has a
diagnosis at end of follow-up}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-get_cohab_corrfac"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-get_cohab_corrfac}{}}}
\subsection{Method \code{get_cohab_corrfac()}}{
Correction factor for cohabitation

Relatives that can be assumed to generally co-habit will be down-weighted,
as some of their similarity in phenotype will be due to shared environment
rather than shared familial genetics. See Kendler et al. 2021b, Supp.Table 2,
Step 3.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$get_cohab_corrfac(rel_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rel_type}}{Character vector, indicating type of relation to proband.
Here, only values \code{M} (mother), \code{F} (father) and \code{SI}
(sibling) will trigger weights different from one}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-get_famsize_corrfac"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-get_famsize_corrfac}{}}}
\subsection{Method \code{get_famsize_corrfac()}}{
Correction factor for family size

Combining liability, age weights, cohabitation correction factor and
degree of relatedness and averaging them across probands for a first
estimate of FGRS will be biased towards larger families. This function
returns a suitable shrinkage factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$get_famsize_corrfac(nrel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nrel}}{Numeric vector containing the weighted number of relatives, where
each relative is multipled by the degree of relatedness to the
proband before summation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-standardize_fgrs"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-standardize_fgrs}{}}}
\subsection{Method \code{standardize_fgrs()}}{
Standardize FGRS value by birth year of proband

The aggregated and shrunk (for family size) FGRS values are additionally
standardized (mean/stddev) per birth year of the proband.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$standardize_fgrs(x, byear)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric, a vector of shrunk FGRS values}

\item{\code{byear}}{numeric, the vector of birth years for the probands}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FGRS_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FGRS_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FGRS_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
