% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{FGRS_phenotypes}
\alias{FGRS_phenotypes}
\title{Population data for disease phenotypes}
\format{
An object of class \code{character} of length 2.
}
\usage{
FGRS_phenotypes
}
\description{
The FGRS package provides the background population data required for
calculating the FGRS for some psychiatric phenotypes in the Swedish
general population. \code{FGRS_phenotypes} is a named vector that lists the
names of the phenotypes currently included in the package. For each
phenotype, the package contains an object of class \code{FGRS_data} that
contains all necessary information; these data objects have the naming
convention \verb{<code>popdata}, where the \verb{<code>} is the 2-4 letter
abbreviation used as name for the phneotype in \code{FGRS_phenotypes}, see
Examples.
}
\examples{
# These are the supported phenotypes
FGRS_phenotypes

# These are the corresponding codes
names(FGRS_phenotypes)

# These are two supported phenotypes
SZpopdat
BDpopdat
}
\references{
Kendler, K., Ohlsson, H., Sundquist, J., & Sundquist, K. (2021b).
Impact of comorbidity on family genetic risk profiles for psychiatric and
substance use disorders: A descriptive analysis. Psychological Medicine, 1-10.
doi:10.1017/S0033291721004268
}
\seealso{
\code{\link[FGRS]{FGRS_data}}
}
\keyword{datasets}
