% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{Example proband and relative data}
\format{
This is a list with two entries, \code{probands} and \code{relatives}. These
entries are data frames with columns as required by function \code{FGRS}
(and documented there).
}
\usage{
ex1
}
\description{
This is a very small example of probands (for whom we want to calculate
an FGRS) and their relatives (whose information is used to calculate
the FGRS).
}
\examples{
ex1

FGRS(ex1$probands, ex1$relatives, phenotype = SZpopdata)
}
\seealso{
\code{\link[FGRS]{FGRS}}
}
\keyword{datasets}
