% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R, R/FGRS.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{Example proband and relative data}
\format{
This is a list with two entries, \code{probands} and \code{relatives}. These
entries are data frames with columns as required by function \code{FGRS}
(and documented there).

A list with two data frames, \code{probands} and \code{relatives}
}
\usage{
ex1

ex1
}
\description{
This is a very small example of probands (for whom we want to calculate
an FGRS) and their relatives (whose information is used to calculate
the FGRS).

A small synthetic data set that tracks the same diagnosis in both
probands and relatives
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{FIXME
}\if{html}{\out{</div>}}
}
\examples{
ex1

FGRS(ex1$probands, ex1$relatives, phenotype = SZpopdata)
}
\seealso{
\code{\link[FGRS]{FGRS}}
}
\keyword{datasets}
