% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGRS.R
\name{FGRS}
\alias{FGRS}
\title{Calculate the family genetic risk score}
\usage{
FGRS(probands, relatives, proband_diag = "SZ", relatives_diag = "SZ")
}
\arguments{
\item{probands}{Data frame containing the proband information; see example
data set \code{ex1$probands} for format.}

\item{relatives}{Data frame containing information on relatives; see example
data set \code{ex1$realtives} for format.}

\item{proband_diag}{Diagnosis for the probands; currently, only \code{SZ} is valid.}

\item{relatives_diag}{Diagnosis for the relatives; currently, only \code{SZ} is valid.}
}
\value{
A list with two elements, \code{probands} and \code{relatives}, data frames
containing the input data and augmented by FGRS (\code{probands}) and
the components for calculating the FGRS (\code{relatives})
}
\description{
Given information on persons of interest (probands) and their relatives,
calculate the family genetic risk score (FGRS).
}
\details{
Still early prototype, format of data / results may change.
}
\examples{
FGRS(ex1$probands, ex1$relatives)
}
